Add Python script to write error message to stderr
This commit adds a new Python script that writes an error message to stderr. The script includes a shebang line and imports the sys module to access the stderr stream. The error message is written using the sys.stderr.write() function. The script also exits with a non-zero status code using sys.exit().