feat: Add script to write contents to a file
This commit adds a new script, 1-writeme.js, which writes the specified contents to a file specified as a command line argument. The script uses the fs module's `writeFile` function to write the contents asynchronously. If an error occurs during the writing process, it will be logged to the console.