feat: Add script to fetch Star Wars character names from a movie
This commit adds a new script, 100-starwars_characters.js, which fetches the names of the Star Wars characters from a specific movie. The script makes an HTTP GET request to the SWAPI API to retrieve the movie data. It then iterates over the character URLs and makes additional requests to fetch each character's data. Finally, it logs the names of the characters to the console.