feat: Add click event to change color of header
This commit adds a new script, 2-script.js, which adds a click event to the element with the ID "red_header". When clicked, it changes the color of the "HEADER" element to red (#FF0000). This enhancement improves the interactivity of the webpage and provides a visual feedback to the user.