feat: Add click event to toggle header color
This commit adds a new script, 4-script.js, which adds a click event to the element with the ID "toggle_header". When clicked, it toggles the color of the "HEADER" element between green and red. This enhancement improves the interactivity of the webpage and provides a visual feedback to the user.