feat: Add click event to add item to list
This commit adds a new script, 5-script.js, which adds a click event to the element with the ID "add_item". When clicked, it appends a new list item with the text "Item" to the unordered list with the class "my_list". This enhancement improves the functionality of the webpage by allowing users to dynamically add items to the list.