feat: Add click event to update header text
This commit adds a new script, 6-script.js, which adds a click event to the element with the ID "update_header". When clicked, it updates the text of the "HEADER" element to "New Header!!!". This enhancement improves the interactivity of the webpage by allowing users to dynamically change the header text.