feat: Add AJAX request to fetch character name from SWAPI API
This commit adds a new script, 7-script.js, which makes an AJAX GET request to the SWAPI API to fetch the name of a character. The retrieved name is then displayed in the element with the ID "character". This enhancement improves the webpage by dynamically updating the character name based on the API response.