feat: Add AJAX request to fetch movies from SWAPI API
This commit adds a new script, 8-script.js, which makes an AJAX GET request to the SWAPI API to fetch a list of movies. The retrieved movie titles are then appended to the unordered list with the ID "list_movies". This enhancement improves the webpage by dynamically populating the movie list based on the API response.