feat: Add AJAX request to fetch French greeting from external API
This commit adds a new script, 9-script.js, which makes an AJAX GET request to an external API (https://hellosalut.stefanbohacek.dev/?lang=fr) to fetch a French greeting. The retrieved greeting is then displayed in the element with the ID "hello". This enhancement improves the webpage by dynamically updating the greeting based on the API response.